requirements.txt
# Hugging Face Spaces Requirements
# Optimized Deepfake Detection System

# Core ML
torch==2.0.1
torchvision==0.15.2
timm==0.9.12
transformers==4.36.0

# Detection
facenet-pytorch==2.5.3
opencv-python-headless==4.8.1.78
Pillow==10.2.0

# Gradio
gradio==4.8.0

# Utilities
numpy==1.24.3
scikit-learn==1.3.2
